processor/pyproject.toml
[project]
name = "processor"
version = "0.1.0"
description = "RAG processor: chunking, embedding, and LanceDB loading"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["rag", "chunking", "embedding", "lancedb", "vector-database"]

dependencies = [
    # CLI & Output
    "click>=8.1.0",
    "rich>=13.0.0",

    # Config
    "pyyaml>=6.0",
    "pydantic>=2.0",

    # Chunking (LlamaIndex + tree-sitter for AST-based code chunking)
    "llama-index-core>=0.10.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.10.0",

    # Database
    "lancedb>=0.4.0",
    "pyarrow>=14.0.0",

    # HTTP (fallback for API calls)
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
# Transformers backend - for jina-code, Qwen3-Embedding, stella via HuggingFace
transformers = [
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "huggingface-hub>=0.20.0",
]

# Multimodal backend - for CLIP/SigLIP image+text embeddings
multimodal = [
    "open-clip-torch>=2.24.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
]

# All GPU backends (transformers + multimodal)
gpu = [
    "processor[transformers,multimodal]",
]

dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

# MCP servers
mcp = [
    "mcp>=1.2.0",
]

# Claude Agent SDK for HyDE optimization (default backend)
claude-sdk = [
    "claude-agent-sdk>=0.1.0",
]

# Cross-encoder reranking for RAG
reranker = [
    "sentence-transformers>=3.0.0",
]

# Everything
all = ["processor[gpu,dev,mcp,reranker,claude-sdk]"]

[project.scripts]
processor = "processor.cli:main"
processor-mcp = "processor_mcp.server:main"
rag-mcp = "rag_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/processor", "mcp/processor_mcp", "mcp/rag_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "ollama: requires running Ollama server",
    "slow: marks tests as slow running",
]
